# packages/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Jomboy Lavka API
  version: 1.0.0
  description: Unified API for Jomboy Lavka client applications
servers:
  - url: https://api.jomboy-lavka.uz/api/v1
    description: Production
  - url: http://localhost:4010/api/v1
    description: Mock server
security:
  - bearerAuth: []
paths:
  /auth/otp/send:
    post:
      tags: [Auth]
      summary: Send OTP to phone
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  example: "+998901234567"
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and get JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, code]
              properties:
                session_id:
                  type: string
                code:
                  type: string
      responses:
        '200':
          description: JWT tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      role:
                        type: string
                      name:
                        type: string
...
  /orders/{order_id}/report-problem:
    post:
      tags: [Order]
      summary: Report problem (creates support ticket)
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [problem_type, description]
              properties:
                problem_type:
                  type: string
                  enum: [not_delivered, damaged, wrong_item, missing_items, other]
                description:
                  type: string
                product_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                photo_urls:
                  type: array
                  items:
                    type: string
                    format: uri
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [open]
  /orders/{order_id}/rate:
    post:
      tags: [Order]
      summary: Rate delivered order
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        '200':
          description: Rating saved
...
  /courier/offers/{order_id}/accept:
    post:
      tags: [Courier]
      summary: Accept delivery offer
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Accepted
        '409':
          description: Offer expired
  /courier/offers/{order_id}/skip:
    post:
      tags: [Courier]
      summary: Skip delivery offer
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Skipped
  /courier/shift:
    post:
      tags: [Courier]
      summary: Start or stop shift
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [start, stop]
                vehicle_type:
                  type: string
                  enum: [foot, bicycle, moped, car]
      responses:
        '200':
          description: Shift updated
...
  /courier/orders/{order_id}/status/pickup:
    post:
      tags: [Courier]
      summary: Picked up at darkstore
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                package_qr:
                  type: string
      responses:
        '200':
          description: Pickup confirmed
  /courier/orders/{order_id}/status/arrived:
    post:
      tags: [Courier]
      summary: Arrived at customer door
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Arrived
...
  /courier/orders/{order_id}/problem:
    post:
      tags: [Courier]
      summary: Report delivery problem
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [address_not_found, customer_unavailable, damage]
                description:
                  type: string
      responses:
        '200':
          description: Problem reported
  /courier/location:
    post:
      tags: [Courier]
      summary: Batch location update
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                points:
                  type: array
                  items:
                    type: object
                    properties:
                      lat:
                        type: number
                      lng:
                        type: number
                      timestamp:
                        type: string
                        format: date-time
      responses:
        '200':
          description: Location saved
...
  /courier/iot/temperature:
    post:
      tags: [IoT, Courier]
      summary: Courier thermal bag temperature reading
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: string
                device_id:
                  type: string
                temperature_c:
                  type: number
                threshold_c:
                  type: number
      responses:
        '200':
          description: Reading recorded
  /orders/{order_id}/timeline:
    get:
      tags: [Support, Phase3]
      summary: Order time machine (Phase 3)
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Timeline events
...
  /customers/{customer_id}/fraud-profile:
    get:
      tags: [Support, Phase3, Antifraud]
      summary: Customer fraud risk profile
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Risk score and flags
...
  /loyalty/promocode/validate:
    post:
      tags: [Loyalty, Phase3]
      summary: Validate promocode
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                subtotal:
                  type: integer
      responses:
        '200':
          description: Validation result
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderId:
      name: order_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...